Some fixes made

+++ src/helpers/stats/stats.h
@@ -59,7 +59,7 @@ double my_t2 = omp_get_wtime();                                                 
 double my_time = (my_t2 - my_t1);                                                           \
 double my_perf = 0;                                        \
 double my_bw = 0;                                  \
-double my_nvals = 0;                               \
+long long int my_nvals = 0;                               \
 FILE *my_f = fopen("perf_stats.txt", "a");                                                 \
 fprintf(my_f, "%s %lf (s) %lf (GFLOP/s) %lf (GB/s) %lld\n", op_name, my_time, my_perf, my_bw, my_nvals);\
 fclose(my_f);
